Settings screen skeleton
- update slidingTabLayout to allow 2 tabs to span the full width of the screen
- add settings activity
- add account and settings fragment
- encapsulate drawer nav with tabbar into new base activity
