+++ app/build.gradle
@@ -27,4 +27,5 @@ dependencies {
     compile 'com.squareup.picasso:picasso:2.4.0'
     compile 'de.hdodenhof:circleimageview:1.2.1'
     compile 'uk.co.chrisjenx:calligraphy:2.0.0'
+    compile project (':library:android-support-v4-preferencefragment')
 }
