+++ app/src/main/java/com/kanshu/kanshu/DevelopmentActivity.java
@@ -4,6 +4,8 @@ import android.content.Intent;
 import android.os.Bundle;
 import android.view.View;
 
+import com.kanshu.kanshu.activity.SettingsActivity;
+
 
 public class DevelopmentActivity extends BaseActivity {
 
@@ -35,6 +37,10 @@ public class DevelopmentActivity extends BaseActivity {
 
     public void onUserMetricsActivity(View view) {
         startActivity(new Intent(this, UserMetricsActivity.class));
+    }
+
+    public void onSettingsActivity(View view) {
+        startActivity(new Intent(this, SettingsActivity.class));
     }
 
 }
