+++ app/src/main/java/com/kanshu/kanshu/fragment/MyAccountFragment.java
@@ -1,0 +1,21 @@
+package com.kanshu.kanshu.fragment;
+
+import android.os.Bundle;
+import android.support.annotation.Nullable;
+import android.support.v4.app.Fragment;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+
+import com.kanshu.kanshu.R;
+
+/**
+ * Created by victorsima on 2/4/15.
+ */
+public class MyAccountFragment extends Fragment {
+
+    @Override
+    public View onCreateView(LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
+        return inflater.inflate(R.layout.fragment_my_account, container, false);
+    }
+}
