+++ app/src/main/res/values/color.xml
@@ -7,4 +7,6 @@
     <color name="article_dart_grey">#606060</color>
     <color name="white">#FFFFFF</color>
     <color name="navbar_off_white">#FAFAFA</color>
+
+    <color name="preferences_divider">#9F9F9F</color>
 </resources>
