+++ app/src/main/res/xml/preferences.xml
@@ -1,0 +1,46 @@
+<?xml version="1.0" encoding="utf-8"?>
+<PreferenceScreen xmlns:android="http://schemas.android.com/apk/res/android">
+
+    <Preference
+        android:title="REMOVE ADS"
+        android:key=""
+        android:layout="@layout/preference_layout"/>
+    <Preference
+        android:title="CHANGE EMAIL"
+        android:key=""
+        android:layout="@layout/preference_layout"/>
+    <Preference
+        android:title="CHANGE PASSWORD"
+        android:key=""
+        android:layout="@layout/preference_layout"/>
+    <Preference
+        android:title="LINKED ACCOUNTS"
+        android:key=""
+        android:layout="@layout/preference_layout"/>
+
+    <PreferenceCategory
+        android:title="ACTIVE FONTS"
+        android:key=""
+        android:layout="@layout/preference_category">
+        <Preference
+            android:title="Font 1"
+            android:key=""
+            android:layout="@layout/preference_layout"/>
+    </PreferenceCategory>
+
+    <PreferenceCategory
+        android:title="SUPPORT"
+        android:key=""
+        android:layout="@layout/preference_category">
+        <Preference
+            android:title="SEND FEEDBACK"
+            android:key=""
+            android:layout="@layout/preference_layout"/>
+
+        <Preference
+            android:title="DELETE ACCOUNT"
+            android:key=""
+            android:layout="@layout/preference_layout"/>
+    </PreferenceCategory>
+
+</PreferenceScreen>
