+++ settings.gradle
@@ -1,1 +1,2 @@
 include ':app'
+include ':library:android-support-v4-preferencefragment'
